Refactor: move ExtractTraceID functions to tracing package
