+++ middleware/http_tracing.go
@@ -6,8 +6,6 @@ import (
 
 	"github.com/opentracing-contrib/go-stdlib/nethttp"
 	"github.com/opentracing/opentracing-go"
-	jaeger "github.com/uber/jaeger-client-go"
-	"golang.org/x/net/context"
 )
 
 // Dummy dependency to enforce that we have a nethttp version newer
@@ -39,33 +37,4 @@ func (t Tracer) Wrap(next http.Handler) http.Handler {
 	}
 
 	return nethttp.Middleware(opentracing.GlobalTracer(), next, options...)
-}
-
-// ExtractTraceID extracts the trace id, if any from the context.
-func ExtractTraceID(ctx context.Context) (string, bool) {
-	sp := opentracing.SpanFromContext(ctx)
-	if sp == nil {
-		return "", false
-	}
-	sctx, ok := sp.Context().(jaeger.SpanContext)
-	if !ok {
-		return "", false
-	}
-
-	return sctx.TraceID().String(), true
-}
-
-// ExtractSampledTraceID works like ExtractTraceID but the returned bool is only
-// true if the returned trace id is sampled.
-func ExtractSampledTraceID(ctx context.Context) (string, bool) {
-	sp := opentracing.SpanFromContext(ctx)
-	if sp == nil {
-		return "", false
-	}
-	sctx, ok := sp.Context().(jaeger.SpanContext)
-	if !ok {
-		return "", false
-	}
-
-	return sctx.TraceID().String(), sctx.IsSampled()
 }
